+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Names preceded by an @ are GitHub usernames.
 ## NEXT
 
 - `.ipa` files can now be opened directly, you don't need to extract the `.app` first. (@DCNick3)
+- New command-line options `--landscape-left` and `--landscape-right` let you change the initial orientation of the device. (@hikari-no-yume)
 
 ## v0.1.0